Fix 'solo monitor' reconnect.
On my Linux system (and maybe others as well), /dev/ttyACM0 is only
created after the Solo key is inserted. If you try to 'monitor' before
inserting, it will fail because the device node does not yet exist. Of
course one can use debug-2 to stall the Solo key until monitor opens the
serial device. However, this has the side effect of making the Solo key
not work without a running monitor.

Second, if you disconnect the Solo key, 'monitor' tries to reconnect,
but does not close the serial device before. The device node might be
deleted, but the minor number is still in use. Therefore, on next
connect e.g. /dev/ttyACM1 will be used and the reconnection will fail.

This change

1. enters reconnect mode also when the device node does not exist
initially

2. closes the device before entering reconnect mode if it was opened
before

(This has been tested on Linux, namely Ubuntu 20.04, and Windows 10. On
Windows it worked fine also without the change.)