Ignore ValueError from fido2 ParseUevent.
The fido2 ParseUevent() method fails if it hits a line missing the '='
in the uevent sysfs file.

As solo-python already live patches the method, add a try-except to the
call of the original method and ignore the ValueError thrown in such a
situation. The same is applied to split() call in this method.

Co-authored-by: Kabbone <58652596+Kabbone@users.noreply.github.com>